requirements.txt
pygame>=2.5.2
requests>=2.31.0
cairosvg>=2.5.0
Pillow>=9.5.0
numpy>=1.24.0
flask>=3.0.0
flask-socketio>=5.3.6
gunicorn>=21.2.0
eventlet>=0.33.3
